저장소 관련 코드 정리
